create table if it doesnt exists

+++ model/DAL/UserDAL.php
@@ -9,13 +9,7 @@ class UserDAL {
         $this->database = $database;
     }
 
-    public function registerUser(\Model\User $user) {
-        $this->database->createTable();
-
-        $username = $user->getUsername();
-        $password = password_hash($user->getPassword(), PASSWORD_BCRYPT);
-
-        // Create connection
+    public function createTableIfNotExists() {
         $connection = new \mysqli(
             $this->database->getHostname(),
             $this->database->getUsername(),
@@ -23,21 +17,59 @@ class UserDAL {
             $this->database->getDatabase()
         );
 
-        // Check connection
-        if ($connection->connect_error) {
-            die("Connection failed: " . $connection->connect_error);
+        if (isset($_SERVER, $_SERVER['HTTP_HOST']) && $_SERVER['HTTP_HOST'] == 'localhost') {
+            if ($connection->connect_error) {
+                die("Connection failed: " . $connection->connect_error);
+            }
         }
 
-        $sql = "INSERT INTO users (username, password)
+        $sql = "CREATE TABLE IF NOT EXISTS users (
+        id INT(6) UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+        username VARCHAR(30) NOT NULL,
+        password VARCHAR(250) NOT NULL,
+        reg_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
+        )";
+
+        if (isset($_SERVER, $_SERVER['HTTP_HOST']) && $_SERVER['HTTP_HOST'] == 'localhost') {
+            if ($connection->query($sql) === TRUE) {
+                return true;
+            } else {
+                echo "Error creating table: " . $connection->error;
+                return false;
+            }
+        }
+    }
+
+    public function registerUser(\Model\User $user) {
+        if ($this->createTableIfNotExists()) {
+
+            $username = $user->getUsername();
+            $password = password_hash($user->getPassword(), PASSWORD_BCRYPT);
+
+            // Create connection
+            $connection = new \mysqli(
+                $this->database->getHostname(),
+                $this->database->getUsername(),
+                $this->database->getPassword(),
+                $this->database->getDatabase()
+            );
+
+            // Check connection
+            if ($connection->connect_error) {
+                die("Connection failed: " . $connection->connect_error);
+            }
+
+            $sql = "INSERT INTO users (username, password)
                 VALUES ('" . $username . "', '" . $password . "')";
 
-        if ($connection->query($sql) === TRUE) {
-            echo "New record created successfully";
-        } else {
-            echo "Error: " . $sql . "<br>" . $connection->error;
-        }
+            if ($connection->query($sql) === TRUE) {
+                echo "New record created successfully";
+            } else {
+                echo "Error: " . $sql . "<br>" . $connection->error;
+            }
 
-        $connection->close();
+            $connection->close();
+        }
     }
 
     public function findExistingUser(\Model\Credentials $credentials) {
